move CORS_list.go and the endpoint to api/
