+++ api/CORS_list.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-package shared
+package api
 
 // CORSList is a list of trusted third-party origins.
 var CORSList = []string{"https://jgraham.github.io"}
