+++ shared/github_oauth.go
@@ -25,8 +25,8 @@ func init() {
 
 // User represents an authenticated GitHub user.
 type User struct {
-	GitHubHandle string
-	GithuhEmail  string
+	GitHubHandle string `json:"handle,omitempty"`
+	GithuhEmail  string `json:"email,omitempty"`
 }
 
 // GitHubAccessControl encapsulates implementation details of access control for the wpt-metadata repository.
